no need to keep the starter script around
When java get's started we don't need the original script anymore, as
it's last action is starting java and no further error checking is
required.

Using exec replaces the script directly, so an external script can start
the launcher script, and reuse it's PID for tracking the running jameica
instance.